+++ packages/app/components/Typo/FooterSpace.tsx
@@ -1,0 +1,5 @@
+import { View } from "native-base";
+
+export const FooterSpace = () => {
+  return <View h={24} />;
+};